Refactored demo
