+++ demo/src/InputFormik.js
@@ -1,21 +1,17 @@
 import React, {useCallback} from 'react';
 import PropTypes from 'prop-types';
 
-import { FormSelect, FormGroup } from 'shards-react';
-import { FormikPropTypes, OptionPropTypes } from '../../../src';
+import {FormGroup, FormInput} from 'shards-react';
 
-const renderOptions = options => options.map(({ value, label }, index) => (
-	<option key={index} value={value}>{label}</option>
-));
+import { FormikPropTypes } from '../../src';
 
-const SelectFormik = ({
+const InputFormik = ({ 
 	name,
 	label,
 	value,
 	error,
 	touched,
 	formik,
-	options,
 	...rest
 }) => {
 	const handleChange = useCallback(event => {
@@ -30,12 +26,10 @@ const SelectFormik = ({
 	const isError = touched && !!error;
 	const isValid = touched && !error;
 
-	const optionsWithNull = [{ label: '', value: '' }].concat(options);
-
 	return (
 		<FormGroup>
 			<label htmlFor={name}>{label}</label>
-			<FormSelect
+			<FormInput 
 				id={name}
 				onChange={handleChange}
 				onBlur={handleBlur}
@@ -43,22 +37,19 @@ const SelectFormik = ({
 				valid={isValid}
 				invalid={isError}
 				{...rest}
-			>
-				{renderOptions(optionsWithNull)}
-			</FormSelect>
+			/>
 			{ isError && <div className="invalid-feedback">{error}</div> }
 		</FormGroup>
 	);
 };
 
-SelectFormik.propTypes = {
+InputFormik.propTypes = {
 	name: PropTypes.string.isRequired,
 	label: PropTypes.string.isRequired,
 	value: PropTypes.string.isRequired,
 	error: PropTypes.string,
 	touched: PropTypes.bool,
 	formik: FormikPropTypes,
-	options: PropTypes.arrayOf(OptionPropTypes),
 };
 
-export default SelectFormik;
+export default InputFormik;
