+++ demo/src/SignupForm/helpers.js
@@ -1,8 +1,3 @@
-export const fetchGhibliMovies = () =>
-	fetch('https://ghibliapi.herokuapp.com/films/')
-		.then(response => response.json())
-		.then(json => json.map(({ title }) => ({ value: title, label: title })));
-
 export const capitals = {
 	'': [],
 	pl: ['Warsaw', 'Warszawa'],
