+++ demo/src/index.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import {render} from 'react-dom';
 
 import { 
@@ -10,7 +10,7 @@ import {
 	Container,
 } from 'shards-react';
 import SignupForm from './SignupForm/SignupForm';
-import SomethingElse from './SomethingElse';
+import BasicForm from './BasicForm';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'shards-ui/dist/css/shards.min.css';
@@ -22,21 +22,28 @@ const Demo = () => {
 				<NavbarBrand href="#">Formuj Demo</NavbarBrand>
 				<Nav navbar>
 					<NavItem>
-						<NavLink href="#">
-							Signup Form
+						<NavLink href="#basic-form">
+							Basic Form
 						</NavLink>
 					</NavItem>
 
 					<NavItem>
-						<NavLink href="#">
-							Something Else
+						<NavLink href="#signup-form">
+							Signup Form
 						</NavLink>
 					</NavItem>
+
 				</Nav>
 			</Navbar>
 
-			<Container style={{padding: '2rem', background: '#eee', height: '100%'}}>
-				<SomethingElse />
+			<Container style={{padding: '2rem', background: '#eee', height: '100%'}} id="basic-form">
+				<h2>Basic Form</h2>
+				<BasicForm />
+			</Container>
+
+			<Container style={{padding: '2rem', background: '#eee', height: '100%'}} id="signup-form">
+				<h2>Signup Form</h2>
+				<SignupForm/>
 			</Container>
 		</div>
 	);
