+++ package.json
@@ -21,7 +21,8 @@
     "test:watch": "jest --watch --verbose"
   },
   "dependencies": {
-    "formik": "^2.1.4"
+    "formik": "^2.1.4",
+    "lodash.merge": "^4.6.2"
   },
   "peerDependencies": {
     "react": "16.x"
@@ -53,7 +54,9 @@
     "form framework"
   ],
   "jest": {
-		"testMatch": ["**/?(*.)+(spec|test).[jt]s?(x)"],
+    "testMatch": [
+      "**/?(*.)+(spec|test).[jt]s?(x)"
+    ],
     "setupFilesAfterEnv": [
       "<rootDir>/setupTests.js"
     ]
